\name{binned_multiSeries}
\alias{binned_multiSeries}
\docType{data}
\title{
Data for vignette example 1. 
}
\description{
A simulated dataseries spanning chr1,chrX,chrY of the human genome (build hg18). Values consist of five series constructed to show mostly random behaviour with the exception of elevated signal in a few regions.
The chromosomes are tiled in 1Mb bins and coordinates are one-based.
}
\usage{data(binned_multiSeries)}
\format{
}
\details{
}
\source{
Simulated data, generated by Shraddha Pai
}
\references{
}
\examples{
data(binned_multiSeries)
head(binned_multiSeries)
}
\keyword{datasets}
