\name{GSM733664_broadPeaks}
\alias{GSM733664_broadPeaks}
\docType{data}
\title{
Data for vignette example 4. 
}
\description{
Broadpeaks file mapping H3K9me3 marks in human lymphoblastoid cells (peaks from chr1, chr2, and chrX).
}
\usage{data(GSM733664_broadPeaks)}
\format{
}
\details{
GEO accession GSM733664, subset containing chr1,chr2,and chrX peaks.
}
\source{

}
\references{
ENCODE Project Consortium, Bernstein BE, Birney E, Dunham I, Green ED, Gunter C, Snyder M. An integrated encyclopedia of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74. 
}
\examples{
data(GSM733664_broadPeaks)
head(GSM733664_broadPeaks)
}
\keyword{datasets}
