\name{hg18_ideo}
\alias{hg18_ideo}
\docType{data}
\title{
Ideogram table for hg18
}
\description{
Cytoband information for all chromosomes in human genome build hg18. Used for vignette examples.
}
\usage{data(hg18_ideo)}
\format{
}
\details{
}
\source{
UCSC genome browser.
}
\references{
}
\examples{
data(hg18_ideo)
head(hg18_ideo)
}
\keyword{datasets}
