% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromWithGenes.R
\name{plotChromosomeWithGenes}
\alias{plotChromosomeWithGenes}
\title{Plot chromosome and genes}
\usage{
plotChromosomeWithGenes(chrom, ideo, bpLim = NULL, titleVertical = 1,
  titleHorizontal = 4, vertical = TRUE, addScale = TRUE,
  chromName_cex = 0.8, verbose = FALSE, GeneList = NULL, ...)
}
\description{
Plot chromosome and genes
}
