\name{wins}
\alias{wins}
\docType{data}
\title{
Data for vignette example 1. 
}
\description{
A simulated dataseries spanning three chromosomes, and containing five series. 
The chromosomes are tiled in 1Mb windows.
}
\usage{data(wins)}
\format{
}
\details{
}
\source{
Simulation by Shraddha Pai
}
\references{
}
\examples{
data(wins)
head(wins)
}
\keyword{datasets}
